Ensure only delete own user, comments, fixed errors

+++ server.js
@@ -27,22 +27,23 @@ app.route('/').get(function(req, res) {
 });
 
 app.route('/signin')
+    // Sign in the user
     .post(function(req, res, next) {
         passport.authenticate('local', function(err, user, info) {
             if (err) {
-                return next(err);
+                // 500 (Internal Server Error)
+                return res.send(500, err);
             }
 
             if (!user) {
-                req.session.messages = [info.message];
-
                 // Credentials not correct, send 401 (Unauthorized)
-                return res.send(401);
+                return res.send(401, info);
             }
 
             req.logIn(user, function(err) {
                 if (err) {
-                    return next(err);
+                    // 500 (Internal Server Error)
+                    return res.send(500, err);
                 }
 
                 res.send(200);
@@ -59,6 +60,7 @@ app.route('/session')
     });
 
 app.route('/users')
+    // Get a list of all the users
     .get(auth.ensureAuthenticated, function(req, res) {
         User.orderBy({
             index: 'username'
@@ -68,16 +70,18 @@ app.route('/users')
             res.json(err);
         });
     })
+    // Create a new user
     .post(function(req, res) {
         User.run().then(function(users) {
             var i = users.length,
                 user;
 
+            // Check if any user already has that username
             while (i--) {
                 if (req.body.username === users[i].username) {
-                    console.log('User %s %s was NOT created', req.body.firstName, req.body.lastName);
+                    console.log('Username "%s" already exits, %s %s', req.body.userame, req.body.firstName, req.body.lastName);
 
-                    // Username aleady exits, send 409 (Conflict)
+                    // Username already exits, send 409 (Conflict)
                     res.status(409);
                     return res.send({
                         message: 'A user with that username already exists'
@@ -92,10 +96,10 @@ app.route('/users')
                 res.status(201);
                 res.send(user.id);
             }, function(err) {
-                console.log('User %s %s was NOT created', user.firstName, user.lastName);
+                console.log('User %s %s NOT created', user.firstName, user.lastName);
                 console.log(err.toString());
 
-                res.status(401);
+                res.status(400);
                 res.send({
                     message: err.toString()
                 });
@@ -104,14 +108,26 @@ app.route('/users')
     });
 
 app.route('/users/:id')
-    .delete(function(req, res) {
-        User.get(req.params.id).run().then(function(user) {
-            user.delete().then(function(user) {
-                console.log('User %s %s removed from the database.', user.firstName, user.lastName);
-                res.status(204);
-                res.send();
+    .delete(auth.ensureAuthenticated, function(req, res) {
+        var id = req.params.id;
+
+        // Only allow people to delete themselves 
+        if (id === req.user.id) {
+            User.get(req.params.id).run().then(function(user) {
+                user.delete().then(function(user) {
+                    console.log('User %s %s removed from the database.', user.firstName, user.lastName);
+
+                    res.status(204);
+                    res.send();
+                });
             });
-        });
+        } else {
+            // Trying to delete other user, send 401 (Unauthorized)
+            res.status(401);
+            res.send({
+                message: 'Not authorized'
+            });
+        }
     });
 
 app.listen(config.port, function() {
